requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
autofaiss==2.17.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
embedding-reader==1.7.0
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.8.0.post1
filelock==3.15.4
fire==0.5.0
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiter==0.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.43.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.0
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
termcolor==2.4.0
together==1.2.10
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.44.0
xxhash==3.5.0
yarl==1.9.8
